requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow>=10.3.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
django-rest-passwordreset==1.3.0
